Refactor Header component to simplify link rendering
- Removed client-side rendering and `usePathname` integration.
- Replaced custom link rendering with `NavLink` component for cleaner code and improved maintainability.
