+++ components/ui/header.tsx
@@ -1,9 +1,5 @@
-"use client";
-
-import Link from "next/link";
-import { usePathname } from "next/navigation";
 import React, { ReactNode } from "react";
-import { cn } from "@/lib/utils/utils";
+import { NavLink } from "./nav-link";
 
 export type HeaderLink = { href: string; label: string };
 
@@ -15,8 +11,6 @@ interface HeaderProps {
 const defaultLinks: HeaderLink[] = [{ href: "/", label: "Dashboard" }];
 
 export function Header({ links = defaultLinks, logo }: HeaderProps) {
-  const pathname = usePathname();
-
   const Logo = (
     <div className="flex items-center gap-2 opacity-70">
       {logo ?? <span>N</span>}
@@ -35,33 +29,11 @@ export function Header({ links = defaultLinks, logo }: HeaderProps) {
           {Logo}
           <nav className="pointer-events-auto">
             <ul className="flex items-center gap-2">
-              {links.map((link) => {
-                const isActive = pathname === link.href;
-                return (
-                  <li key={link.href}>
-                    {isActive ? (
-                      <span
-                        className="rounded-full px-4 py-1.5 text-sm font-medium backdrop-blur-md bg-white/10 ring-1 ring-white/20 text-zinc-400 cursor-default"
-                        aria-current="page"
-                      >
-                        {link.label}
-                      </span>
-                    ) : (
-                      <Link
-                        href={link.href}
-                        className={cn(
-                          "rounded-full px-4 py-1.5 text-sm font-medium backdrop-blur-md",
-                          "bg-white/5 ring-1 ring-white/10 text-zinc-100/90",
-                          "hover:bg-white/8 hover:ring-white/20 transition",
-                          "outline-none focus-visible:ring-2 focus-visible:ring-emerald-400/60"
-                        )}
-                      >
-                        {link.label}
-                      </Link>
-                    )}
-                  </li>
-                );
-              })}
+              {links.map((link) => (
+                <li key={link.href}>
+                  <NavLink href={link.href}>{link.label}</NavLink>
+                </li>
+              ))}
             </ul>
           </nav>
         </div>
